# backend/Sistema_Crud_API_PIBIC/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SistemaAPI PIBIC
  version: 1.0.0
  description: |
    API REST para gerenciamento e resolucao de casos clinicos academicos.
    As operacoes de IA aceitam somente dados sinteticos ou previamente
    desidentificados e toda saida exige revisao humana antes da publicacao.
    O recurso e educacional e nao substitui decisao clinica.
  contact:
    name: Equipe SistemaAPI PIBIC
  license:
    name: Uso academico
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /usuarios:
    get:
      summary: Lista usuarios de acesso
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Usuarios encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaUsuario"
    post:
      summary: Cadastra usuario com senha BCrypt e role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequest"
      responses:
        "201":
          description: Usuario criado
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /usuarios/{id}:
    get:
      summary: Busca usuario por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Usuario encontrado
    put:
      summary: Atualiza usuario, senha e role
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequest"
      responses:
        "200":
          description: Usuario atualizado
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /usuarios/{id}/ativar:
    patch:
      summary: Ativa usuario
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Usuario ativado
  /usuarios/{id}/desativar:
    patch:
      summary: Desativa usuario
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Usuario desativado
...
  /professores:
    get:
      summary: Lista professores
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Professores encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaProfessor"
    post:
      summary: Cadastra professor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessorRequest"
      responses:
        "201":
          description: Professor criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorResponse"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /professores/cadastro:
    post:
      summary: Cadastra professor e sua credencial de acesso
      description: Operacao restrita a administradores.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessorCadastroRequest"
      responses:
        "201":
          description: Professor e credencial criados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /professores/{id}:
    get:
      summary: Busca professor por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Professor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorResponse"
    put:
      summary: Atualiza professor
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessorRequest"
      responses:
        "200":
          description: Professor atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorResponse"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
    delete:
      summary: Remove professor
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Professor removido
  /professores/{id}/casos:
    get:
      summary: Lista casos de um professor
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Casos do professor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaCasoClinico"
  /professores/{id}/relatorio-desempenho:
    get:
      summary: Gera relatorio de desempenho do professor
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Relatorio gerado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatorioDesempenhoProfessor"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    page:
      name: page
      in: query
      description: Indice da pagina, iniciando em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: Quantidade de elementos por pagina.
      schema:
        type: integer
        minimum: 1
        default: 20
    sort:
      name: sort
      in: query
      description: Ordenacao no formato campo,direcao.
      schema:
        type: string
        example: nome,asc
  responses:
...
    CorpoRequisicaoExcedido:
      description: O corpo da requisicao excede o limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:corpo-requisicao-excedido
            title: Corpo da requisicao excedido
            status: 413
            detail: O corpo da requisicao excede o limite permitido
            instance: /auth/login
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O corpo da requisicao excede o limite permitido
...
    DadosInvalidos:
      description: A requisicao possui campos, JSON, parametros ou tipos invalidos.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NaoAutenticado:
      description: Token ausente, invalido, expirado ou usuario desativado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AcessoNegado:
      description: O usuario autenticado nao possui permissao para a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflito:
      description: O estado atual ou a concorrencia dos dados impede a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos ate a repeticao recomendada quando o conflito for transitorio.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    CorrelationId:
      description: Identificador da requisicao para rastreamento ponta a ponta.
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - correlationId
        - erro
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        erro:
          type: string
          deprecated: true
          description: Alias legado de detail; mantido temporariamente para compatibilidade.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo quando a validacao falha.
        campos:
          type: object
          deprecated: true
          additionalProperties:
            type: string
          description: Alias legado de errors; mantido temporariamente para compatibilidade.
...
    UsuarioRequest:
      type: object
      required:
        - username
        - senha
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        senha:
          type: string
          format: password
          minLength: 12
          maxLength: 72
        role:
          type: string
          enum: [ADMIN, PROFESSOR, ALUNO]
        ativo:
          type: boolean
        idAluno:
          type: integer
          format: int64
          nullable: true
          description: Obrigatorio para usuarios com role ALUNO.
        idProfessor:
          type: integer
          format: int64
          nullable: true
          description: Obrigatorio para usuarios com role PROFESSOR.
    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, PROFESSOR, ALUNO]
        ativo:
          type: boolean
        idAluno:
          type: integer
          format: int64
          nullable: true
        idProfessor:
          type: integer
          format: int64
          nullable: true
...
    ProfessorRequest:
      type: object
      required:
        - nome
        - email
        - materia
      properties:
        nome:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        materia:
          type: string
          maxLength: 120
    ProfessorCadastroRequest:
      allOf:
        - $ref: "#/components/schemas/ProfessorRequest"
        - type: object
          required:
            - username
            - senha
          properties:
            username:
              type: string
              minLength: 3
              maxLength: 100
            senha:
              type: string
              format: password
              minLength: 12
              maxLength: 72
    ProfessorResponse:
      allOf:
        - $ref: "#/components/schemas/ProfessorRequest"
        - type: object
          properties:
            id:
              type: integer
              format: int64
...
    CasoClinicoRequest:
      type: object
      required:
        - titulo
        - disciplina
        - areaSaude
        - estilo
        - especialidade
        - nivelDificuldade
      properties:
        idProfessor:
          type: integer
          format: int64
          nullable: true
        titulo:
          type: string
          maxLength: 200
        disciplina:
          type: string
          maxLength: 120
        areaSaude:
          type: string
          maxLength: 120
        estilo:
          type: string
          maxLength: 60
        especialidade:
          type: string
          maxLength: 120
        objetivoAprendizagem:
          type: string
          maxLength: 10000
        nivelDificuldade:
          type: string
          enum: [BAIXA, MEDIA, ALTA]
        tempoLimiteMinutos:
          type: integer
          minimum: 15
          maximum: 480
          default: 60
          description: Tempo total para responder o caso completo.
    CasoClinicoResponse:
      allOf:
        - $ref: "#/components/schemas/CasoClinicoRequest"
        - type: object
          properties:
            idCaso:
              type: integer
              format: int64
            status:
              type: string
              enum: [RASCUNHO, PUBLICADO, ARQUIVADO]
            dataCriacao:
              type: string
              format: date-time
            dataAtualizacao:
              type: string
              format: date-time
...
    PaginaBase:
      type: object
      required:
        - totalElements
        - totalPages
        - number
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    PaginaProfessor:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/ProfessorResponse"
    PaginaUsuario:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/UsuarioResponse"
...
    PaginaCasoClinico:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/CasoClinicoResponse"
...
    RelatorioDesempenhoProfessor:
      type: object
      properties:
        idProfessor:
          type: integer
          format: int64
        totalRespostas:
          type: integer
          format: int64
        totalAvaliadas:
          type: integer
          format: int64
          description: Respostas automaticas ou revisadas incluidas no calculo do aproveitamento medio.
        totalPendentesRevisao:
          type: integer
          format: int64
        totalCorretas:
          type: integer
          format: int64
        aproveitamentoMedio:
          type: number
          format: double
          description: Percentual calculado somente sobre as respostas ja avaliadas.
...